refactor(metadata): Extract balance alert operations to alert.rs
- Create new alert.rs module with 11 balance alert methods
- Move notification logging, CRUD, state management, and validation methods
- Keep blockchain and exchange rate operations in db.rs
- Add alert module declaration to mod.rs

Closes #58
